Split circuit operations between headers and source
